move dataset from utils namespace to top level
