+++ expertise/models/centroid/setup.py
@@ -6,7 +6,7 @@ import numpy as np
 import openreview
 from expertise.utils.vocab import Vocab
 from expertise.utils.batcher import Batcher
-from expertise.utils.dataset import Dataset
+from expertise.dataset import Dataset
 import expertise.utils as utils
 from expertise.utils.data_to_sample import data_to_sample
