+++ expertise/models/tfidf/infer_tfidf.py
@@ -3,7 +3,7 @@ import csv, json
 from collections import defaultdict
 from expertise import utils
 from expertise.config import Config
-from expertise.utils.dataset import Dataset
+from expertise.dataset import Dataset
 from datetime import datetime
 import multiprocessing as mp
