+++ expertise/utils/standard_test.py
@@ -1,7 +1,7 @@
 import os
 from collections import defaultdict
 from expertise import utils
-from expertise.utils.dataset import Dataset
+from expertise.dataset import Dataset
 from expertise.evaluators.mean_avg_precision import eval_map
 from expertise.evaluators.hits_at_k import eval_hits_at_k
 from tqdm import tqdm
